renamed to artifact-provider
