+++ src/index.mts
@@ -1,12 +1,12 @@
 import type { Config } from './config.mjs';
 import { DebBuilder } from './deb/deb-builder.mjs';
 import type { Deployer } from './deployer.mjs';
-import JfrogArtifactsProvider from './jfrog/artifacts-provider.mjs';
+import JfrogArtifactProvider from './jfrog/artifact-provider.mjs';
 
 export { type Config, createConfigProvider } from './config.mjs';
 
 export function plan(config: Config): Promise<void[]> {
-	const artifactsProvider = new JfrogArtifactsProvider(config);
+	const artifactsProvider = new JfrogArtifactProvider(config);
 
 	const builders: Deployer[] = [];
